Add jumping
I added a box collider to the player and a check to make sure it's on the ground so when you click jump you'll only jump if you're touching the ground. I also played with the physics a little so it feels a little nicer to move around. This should be working with controller and keyboard. I haven't tested with controller but keyboard is working.