Create Simple Dialogue System
I created a simple dialogue system that just goes through an array of text and shows you what it says. There's also this weird... not bug. But something in the text array that doesn't let you see what you're typing in the first element. So I coded the dialogue system to skip the first element in the array so you don't have to worry about it.