Add collectable logic framework
I added the collectable system. Its automatically checks how many collectables are in your scene and then keeps that in the game logic. So whenever you pick up the collectable it removes one from the game logic. Nothing happens yet but I'm hoping to connect this to the dialogue system to have it win