Update the dialogue system
I changed the dialogue system to NOT reset whenever you leave the hitbox and comeback. It takes you to the 2nd to last piece of dialogue and then if you press interact again. It takes you to either the repeating text if you haven't completed his "quest" or to the success text.