Comment code and remove bugs
I commented the code so hopefully it will be easier to understand.  I also added in the next scene and a portal to take you to the next scene and I added in some code to remove some bugs I introduced. I also added in some cool optimization stuff to make the hierarchy better and warning system