Tweak movement
I added changed the movement a little bit so that you can change the speed and jump height to be twice as much. Also, I found this new mechanic where you can store your previous jump speed and add the new jump speed. It's really cool and just works. But with the current movement settings. You can make it to the ledge all the way to the left. I really like the movement