Fix force grenade
I added the functionality for the grenade but after testing. I figured out that it wasn't moving on the X axis and it was because of the way that I was calculating movement. I changed it to be more physics based but I had to do a lot of fiddly things. I had to add a very small friction pad on the player in order to make it not so slippery but still not stick to the walls.