Add features
I updated the UI to include a use button and I also added in another type of grenade. The anchor grenade. It teleports you but it doesn't move from where it was placed. Also updated the logic a bit so that it checks if you won the level instead of if you collected all coins