Update dynamic dialogue
I added more variables that you can add to the dialogue that uses the game to reference the number. So you can just type in the number you want and it will dynamically be generated. for example if you wanted to say. "There are 10 collectables on the map you have collected 3 so far." You would type "There are (current number) collectables on the map you have collected (collected number) so far."