Add level UI
I added the UI and the functionality to move between levels and start a level over again. Everything is working pretty well. But there was a bug that was introduced. I have no clue how to fix it, or why it's throwing an error. The game is working fine though so... I'm just not gonna worry about it.