Add features
I added in an explosion effect to the camera when the blue crystal explodes and I also added in more sound effects and I did some QOL changes to the UI and the way crystals work. Now when you pick up crystals it automatically selects that crystal to use. Hopefully that makes it clear which crystal you're going to use.