Fix timer bug
There was a timer bug that would sometimes return a value of a float instead of an int when rounding. I had to move over to the system math function in order to fix it. I had to convert it from a double to a float in order for it to be usable with unity.