Fix item spawn bug
There was a bug that was introduced that would sometimes make the item spawn position go really far away from the player. I fixed that by making it so if the spawn went farther than 2 units away from the player, it would teleport the item spawn point to the players position.